move all settings classes into 'settings' package
