+++ app/src/main/AndroidManifest.xml
@@ -36,7 +36,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".PreferencesActivity"
+            android:name=".settings.PreferencesActivity"
             android:label="@string/title_activity_settings"/>
         <activity
             android:name=".AcknowledgementsActivity"
